src/thumbnailGenerator/requirements.txt
os
boto3
Pillow